fixed up the pinpoint model and improved its speed to upto 300M ins / second on fast forward mode (instrumented at BB granularity)

+++ src/config.h
@@ -35,7 +35,7 @@ class config {
         MEM_MODEL getMemModel ();
         BP_TYPE getBPtype ();
         int getNumEu ();
-        float getMaxInsCnt ();
+        long double getMaxInsCnt ();
         bool isEnSquash ();
         bool isEnEuFwd ();
         bool isEnMemFwd ();
@@ -61,7 +61,7 @@ class config {
         string _out_path;
         string _bench_path;
         bool _enable_log_stat;
-        float _max_ins_cnt;
+        long double _max_ins_cnt;
 
 		//PinPoint Config
 		char pinPoint_s_file[PARSE_LEN];
